+++ firmware/hdl/packages/gt_mp7_core_pkg.vhd
@@ -3,12 +3,12 @@
 
 -- actual versions:
 -- use "FRAME_VERSION" as mp7_ugt release fw version (used for tag name).
--- mp7_ugt (=FRAME_VERSION): v1.21.7
+-- mp7_ugt (=FRAME_VERSION): v1.22.0
 
 -- use "GT_VERSION" as mp7_ugt release fw version (used for tag name).
--- gt: v1.21.7
+-- gt: v1.22.0
 -- frame: v1.3.1 (see frame.vhd)
--- gtl: v1.17.6 (see gtl_module_tpl.vhd)
+-- gtl: v1.18.0 (see gtl_module_tpl.vhd)
 -- fdl: v1.4.1 (see fdl_module.vhd)
 
 -- HB 2022-11-23: v1.21.6 - Updated doc/scales_inputs_2_ugt (added definition of muon shower bits).
@@ -31,6 +31,7 @@
 -- v1.15.2: Added tcl script for "manualy" bit file generation (after timing errors)
 --
 -- gtl history:
+-- HB 2023-01-31: v1.18.0: Implemented logic for CICADA (bjets).
 -- HB 2022-11-29: v1.17.6: Bug fix in algo_pre_scaler_fractional_float.vhd.
 -- HB 2022-11-16: v1.17.5: Bug fix in correlation_conditions.vhd.
 -- HB 2022-09-23: v1.17.4: Used "delay_pipeline" for condition output in esums_conditions.vhd, min_bias_hf_conditions.vhd and towercount_condition.vhd.
@@ -85,8 +86,8 @@ package gt_mp7_core_pkg is
 -- ==================================================================================================
 -- GT firmware version
     constant GT_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant GT_MINOR_VERSION      : integer range 0 to 255 := 21;
-    constant GT_REV_VERSION        : integer range 0 to 255 := 7;
+    constant GT_MINOR_VERSION      : integer range 0 to 255 := 22;
+    constant GT_REV_VERSION        : integer range 0 to 255 := 0;
 	constant GT_VERSION : std_logic_vector(31 downto 0) := X"00" &
            std_logic_vector(to_unsigned(GT_MAJOR_VERSION, 8)) &
            std_logic_vector(to_unsigned(GT_MINOR_VERSION, 8)) &
@@ -97,8 +98,8 @@ package gt_mp7_core_pkg is
     constant FRAME_REV_VERSION        : integer range 0 to 255 := 1;
 -- GTL firmware (fix part) version
     constant GTL_FW_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant GTL_FW_MINOR_VERSION      : integer range 0 to 255 := 17;
-    constant GTL_FW_REV_VERSION        : integer range 0 to 255 := 6;
+    constant GTL_FW_MINOR_VERSION      : integer range 0 to 255 := 18;
+    constant GTL_FW_REV_VERSION        : integer range 0 to 255 := 0;
 -- FDL firmware version
     constant FDL_FW_MAJOR_VERSION      : integer range 0 to 255 := 1;
     constant FDL_FW_MINOR_VERSION      : integer range 0 to 255 := 4;